feat(PoolAllocator): add PoolAllocator
Add an abstraction for creating a Pool of nodes

Closes: #216
Approved by: jmcadden

+++ src/hosted/NodeAllocator.cc
@@ -185,9 +185,15 @@ ebbrt::NodeAllocator::NodeAllocator() : node_index_(2), allocation_index_(0) {
                         ".IPAM.Config}}{{.Gateway}}{{end}}' " +
                         network_id_);
   } else {
-    std::cerr << "Creating custom network" << std::endl;
-    network_id_ = RunCmd(CustomNetworkCreate);
-    network_ip = RunCmd(CustomNetworkIp + network_id_);
+    if (CustomNetworkNodeArguments.find("weave") != std::string::npos) {
+      std::cerr << "Connecting to weave network" << std::endl;
+      network_id_ = "weave";
+      network_ip = RunCmd(CustomNetworkIp);
+    } else {
+      std::cerr << "Creating custom network" << std::endl;
+      network_id_ = RunCmd(CustomNetworkCreate);
+      network_ip = RunCmd(CustomNetworkIp + network_id_);
+    }
   }
 
   uint8_t ip0, ip1, ip2, ip3;
@@ -248,7 +254,7 @@ void ebbrt::NodeAllocator::AppendArgs(std::string arg) {
 ebbrt::NodeAllocator::NodeDescriptor
 ebbrt::NodeAllocator::AllocateNode(std::string binary_path, int cpus,
                                    int numaNodes, int ram,
-                                   std::string arguments) {
+                                   std::string arguments, std::string constraint_node) {
 
   assert(cpus != 0 && numaNodes != 0 && ram != 0);
 
@@ -274,6 +280,11 @@ ebbrt::NodeAllocator::AllocateNode(std::string binary_path, int cpus,
 #ifndef NDEBUG
   docker_args << " --expose 1234 -e DEBUG=true";
 #endif
+
+  if (!constraint_node.empty()) {
+    docker_args << " -e constraint:node==" << constraint_node << " ";
+  }
+
   if (CustomNetworkNodeArguments.empty()) {
     docker_args << " --net=" << network_id_ << " ";
   } else {
@@ -331,6 +342,10 @@ ebbrt::NodeAllocator::AllocateNode(std::string binary_path, int cpus,
 }
 ebbrt::NodeAllocator::~NodeAllocator() {
   nodes_.clear();
+  /* If weave is the network don't remove it */
+  if (CustomNetworkNodeArguments.find("weave") != std::string::npos) {
+    return;
+  }
   if (CustomNetworkRemove.empty()) {
     std::cerr << "removing Network: " << network_id_.substr(0, 12) << std::endl;
     RunCmd("docker network rm " + network_id_);
